ignore_dir_patterns: prune scan/search tree by entire dirs
watchdog uses Watchers implementing a `should_watch_dir` method
which can be used to prune the tree of files returned by the
scan. If the dir is accepted, the function recurses, scanning the
subdirectory, otherwise, it is skipped (and thus the entire
sub-tree is pruned).

This change accepts regex patterns which, if matching, will cause
a directory (and therefore all its contents) to be skipped.
This is useful for directories such as '.node_modules', '.git' and
so on.