Performance tweak
This change keeps the temporary output file open and reuses it
so long as no file needs to be overwritten.

Since the first compile iteration starts with an empty database of file
hashes, all files would be overwritten by default. To avoid this, we
introduce a boolean to mark wether any snippets have been expanded
during the parsing of a particular file.
If not, the file cannot have changed and so we won't replace it. This
almost halves the execution time in large projects with no templated
files (e.g. the Linux kernel sources)