embed location information in AST
Tokenizer.location() overhauled. It is computed on each call
to next() as we advance. It should be faster than the prior
approach if queried for most tokens.

Tokenizer.location() now returns the START of the token rather
than the END as before. This should be more intuitive in error
reporting.

AST nodes Block, Literal and Expr now contain location information.
This will enable errors raised in the interpreter to point to the
line & column which caused the error.